Add viewmodel provider and clean up package names
Viewmodel provider allows to define and access viewmodel from
FXML despite the absence of default ctor. Package names now follow
Oracle's convention of no uppercase letters.
