+++ code/newton-method/Infrastructure/build.gradle
@@ -4,6 +4,6 @@ dependencies {
 }
 
 sourceSets {
-    main.java.srcDir 'src/main'
-    test.java.srcDir 'src/test'
+    main.java.srcDir 'src/main/java'
+    test.java.srcDir 'src/test/java'
 }
