+++ code/newton-method/Infrastructure/src/test/java/ru/unn/agile/newtonroots/infrastructure/NewtonRootAppLoggingTestsWithPlainTextLogger.java
@@ -1,0 +1,18 @@
+package ru.unn.agile.newtonroots.infrastructure;
+
+import ru.unn.agile.newtonroots.viewmodel.NewtonRootAppLoggingTests;
+import ru.unn.agile.newtonroots.viewmodel.NewtonRootAppViewModel;
+
+import java.io.IOException;
+
+public class NewtonRootAppLoggingTestsWithPlainTextLogger extends NewtonRootAppLoggingTests {
+    @Override
+    public void setUp() {
+        try {
+            PlainTextLogger logger = new PlainTextLogger("newton-method.log");
+            viewModel = new NewtonRootAppViewModel(logger);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
