--- code/newton-method/Model/src/main/java/ru/unn/agile/NewtonRoots/Model/FunctionInterface.java
@@ -1,5 +1,0 @@
-package ru.unn.agile.NewtonRoots.Model;
-
-public interface FunctionInterface {
-    double compute(double x);
-}
