+++ code/newton-method/Model/src/main/java/ru/unn/agile/newtonroots/model/FunctionInterface.java
@@ -1,0 +1,5 @@
+package ru.unn.agile.newtonroots.model;
+
+public interface FunctionInterface {
+    double compute(double x);
+}
