+++ code/newton-method/View/src/main/java/ru/unn/agile/newtonroots/view/Main.java
@@ -1,0 +1,23 @@
+package ru.unn.agile.newtonroots.view;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+public class Main extends Application {
+
+    @Override
+    public void start(final Stage primaryStage) throws Exception {
+        Parent root = FXMLLoader.load(getClass().getResource("NewtonRootsApp.fxml"));
+        primaryStage.setTitle("Root finder");
+        primaryStage.setScene(new Scene(root));
+        primaryStage.setResizable(false);
+        primaryStage.show();
+    }
+
+    public static void main(final String[] args) {
+        launch(args);
+    }
+}
