+++ code/newton-method/View/src/main/java/ru/unn/agile/newtonroots/view/NewtonRootsApp.java
@@ -1,0 +1,46 @@
+package ru.unn.agile.newtonroots.view;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.*;
+import ru.unn.agile.newtonroots.viewmodel.NewtonRootAppViewModel;
+import ru.unn.agile.newtonroots.model.NewtonMethod.StoppingCriterion;
+
+public class NewtonRootsApp  {
+    @FXML
+    private NewtonRootAppViewModelProvider viewModelProvider;
+    @FXML
+    private TextField leftPointText;
+    @FXML
+    private TextField rightPointText;
+    @FXML
+    private TextField derivativeStepText;
+    @FXML
+    private TextField accuracyText;
+    @FXML
+    private TextField functionText;
+    @FXML
+    private TextArea reportTextField;
+    @FXML
+    private Button findRootButton;
+    @FXML
+    private TextField startPointText;
+    @FXML
+    private ChoiceBox<StoppingCriterion> stopCriterionSelector;
+
+    @FXML
+    private void initialize() {
+        NewtonRootAppViewModel viewModel = viewModelProvider.getViewModel();
+
+        leftPointText.textProperty().bindBidirectional(viewModel.leftPointProperty());
+        rightPointText.textProperty().bindBidirectional(viewModel.rightPointProperty());
+        derivativeStepText.textProperty().bindBidirectional(viewModel.derivativeStepProperty());
+        accuracyText.textProperty().bindBidirectional(viewModel.accuracyProperty());
+        functionText.textProperty().bindBidirectional(viewModel.functionProperty());
+        reportTextField.textProperty().bindBidirectional(viewModel.solverReportProperty());
+        startPointText.textProperty().bindBidirectional(viewModel.startPointProperty());
+
+        stopCriterionSelector.valueProperty().bindBidirectional(viewModel.stopCriterionProperty());
+
+        findRootButton.setOnAction(value -> viewModel.findRoot());
+    }
+}
