--- code/newton-method/View/src/main/resources/ru/unn/agile/NewtonRoots/view/NewtonRootsApp.fxml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.control.TitledPane?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import ru.unn.agile.NewtonRoots.viewmodel.*?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="420.0" prefWidth="418.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.unn.agile.NewtonRoots.view.NewtonRootsApp">
-    <fx:define>
-        <NewtonRootAppViewModel fx:id="viewModel" />
-    </fx:define>
-      <Button fx:id="findRootButton" disable="${viewModel.findRootButtonDisable}" layoutX="14.0" layoutY="265.0" mnemonicParsing="false" text="Find root" />
-      <TitledPane animated="false" layoutX="14.0" layoutY="14.0" prefHeight="240.0" prefWidth="398.0" text="Parameters">
-          <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="237.0" prefWidth="396.0">
-                  <Label layoutX="20.0" layoutY="9.0" text="Left point:" />
-                  <Label layoutX="20.0" layoutY="36.0" text="Right point:" />
-                  <Label layoutX="20.0" layoutY="64.0" text="Derivative step:" />
-                  <Label layoutX="20.0" layoutY="93.0" prefWidth="64.0" text="Accuracy:" />
-                  <TextField fx:id="leftPointText" layoutX="132.0" layoutY="4.0" prefHeight="26.0" prefWidth="88.0"/>
-                  <TextField fx:id="rightPointText" layoutX="132.0" layoutY="31.0" prefHeight="26.0" prefWidth="88.0" />
-                  <TextField fx:id="derivativeStepText" layoutX="132.0" layoutY="59.0" prefHeight="26.0" prefWidth="88.0" />
-                  <TextField fx:id="accuracyText" layoutX="132.0" layoutY="88.0" prefHeight="26.0" prefWidth="88.0" />
-                  <Label layoutX="20.0" layoutY="124.0" text="Function:" />
-                  <TextField fx:id="functionText" layoutX="132.0" layoutY="118.0" prefHeight="26.0" prefWidth="176.0" />
-                  <Label layoutX="20.0" layoutY="152.0" text="Start point:" />
-                  <TextField fx:id="startPointText" layoutX="132.0" layoutY="147.0" prefHeight="26.0" prefWidth="88.0" />
-                  <Label layoutX="20.0" layoutY="182.0" text="Stop criterion:" />
-                  <ChoiceBox fx:id="stopCriterionSelector" items="${viewModel.stopCriteria}"
-                             layoutX="132.0" layoutY="177.0" prefWidth="150.0" />
-            </AnchorPane>
-      </TitledPane>
-      <TitledPane animated="false" layoutX="11.0" layoutY="301.0" prefHeight="106.0" prefWidth="404.0" text="Solver report">
-          <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="77.0" prefWidth="402.0">
-                  <TextArea fx:id="reportTextField" editable="false" prefHeight="80.0" prefWidth="404.0" />
-            </AnchorPane>
-      </TitledPane>
-      <Label fx:id="statusLabel" layoutX="102.0" layoutY="270.0" text="${viewModel.applicationStatus}" />
-</AnchorPane>
