+++ code/newton-method/View/src/test/java/ru/unn/agile/newtonroots/view/ViewTests.java
@@ -1,0 +1,14 @@
+package ru.unn.agile.newtonroots.view;
+
+import org.junit.Ignore;
+import org.junit.Test;
+
+import static org.junit.Assert.fail;
+
+public class ViewTests {
+    @Test
+    @Ignore
+    public void fakeTest() {
+        fail("No tests for view");
+    }
+}
