--- code/newton-method/ViewModel/src/main/java/ru/unn/agile/NewtonRoots/viewmodel/Logger.java
@@ -1,9 +1,0 @@
-package ru.unn.agile.NewtonRoots.viewmodel;
-
-import java.util.List;
-
-public interface Logger {
-    List<String> getMessageList();
-    String getLastMessage();
-    void appendMessage(String message);
-}
