+++ code/newton-method/ViewModel/src/main/java/ru/unn/agile/newtonroots/viewmodel/Logger.java
@@ -1,0 +1,9 @@
+package ru.unn.agile.newtonroots.viewmodel;
+
+import java.util.List;
+
+public interface Logger {
+    List<String> getMessageList();
+    String getLastMessage();
+    void appendMessage(String message);
+}
