+++ code/newton-method/ViewModel/src/test/java/ru/unn/agile/newtonroots/viewmodel/FakeLogger.java
@@ -1,0 +1,27 @@
+package ru.unn.agile.newtonroots.viewmodel;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class FakeLogger implements Logger {
+    private final ArrayList<String> messageList = new ArrayList<>();
+
+    @Override
+    public List<String> getMessageList() {
+        return messageList;
+    }
+
+    @Override
+    public String getLastMessage() {
+        if (messageList.isEmpty()) {
+            return null;
+        }
+
+        return messageList.get(messageList.size() - 1);
+    }
+
+    @Override
+    public void appendMessage(String message) {
+        messageList.add(message);
+    }
+}
